% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FnR-package.R
\docType{package}
\name{FnR-package}
\alias{FnR}
\alias{FnR-package}
\title{FnR: Inbreeding and Numerator Relationship Coefficients}
\description{
Compute inbreeding coefficients using the method of Meuwissen and Luo (1992) \doi{ 10.1186/1297-9686-24-4-305}, and numerator relationship coefficients between individuals using the method of Van Vleck (2007) <PMID:18050089>.
}
\author{
\strong{Maintainer}: Mohammad Ali Nilforooshan \email{m.a.nilforooshan@gmail.com} (\href{https://orcid.org/0000-0003-0339-5442}{ORCID})

}
\keyword{internal}
