% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notsire_notdam.R
\name{notsire_notdam}
\alias{notsire_notdam}
\title{Compute numerator relationship coefficients between non-sires and non-dams}
\usage{
notsire_notdam(ped, notsire, notdam)
}
\arguments{
\item{ped}{: A data frame with the first columns being ID, SIRE, and DAM.
Missing parents (SIRE and DAM) are denoted as 0.}

\item{notsire}{: A set of individual IDs that cannot be found among \code{ped$SIRE}.}

\item{notdam}{: A set of individual IDs (distict from \code{notsire}) that cannot be found among \code{ped$DAM}.}
}
\value{
: Numerator relationship coefficients between \code{notsire} and \code{notdam} individuals.
}
\description{
Compute numerator relationship coefficients between non-sires and non-dams
}
\examples{
message("TODO:")

}
