% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sire_sire.R
\name{sire_sire}
\alias{sire_sire}
\title{Compute numerator relationship coefficients between two groups of sires}
\usage{
sire_sire(ped, set1, set2)
}
\arguments{
\item{ped}{: A data frame with the first columns being ID, SIRE, and DAM.
Missing parents (SIRE and DAM) are denoted as 0.}

\item{set1}{: A set of sire IDs.}

\item{set2}{: A set of sire IDs, distict from \code{set1}.}
}
\value{
: Numerator relationship coefficients between \code{set1} and \code{set2} individuals.
}
\description{
Compute numerator relationship coefficients between two groups of sires
}
\examples{
message("TODO:")

}
