% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dam_dam.R
\name{dam_dam}
\alias{dam_dam}
\title{Compute numerator relationship coefficients between two groups of dams}
\usage{
dam_dam(ped, set1, set2)
}
\arguments{
\item{ped}{: A data frame with integer columns corresponding to ID, SIRE, and DAM. IDs should be sequential, starting from 1. Missing parents (SIRE and DAM) are denoted as 0.}

\item{set1}{: A set of dam IDs.}

\item{set2}{: A set of dam IDs, distinct from \code{set1}.}
}
\value{
: Numerator relationship coefficients between \code{set1} and \code{set2} individuals in the form of a matrix.
}
\description{
Compute numerator relationship coefficients between two groups of dams
}
\examples{
message("TODO: Write example(s).")

}
